fix: incorrect logic for distance between cities, slow queries
Previously, getting the nearest city or recommendations could take >15 seconds. Now it typically takes 20-30 ms.
This is achieved through getting rid of a nested loop and using a faster but less accurate distance algorithm.
The difference in distance can be over 30km, but that's not significant for our uses, with less than 300 possible cities to match with worldwide.